Update Space.cs
Change used to long #3

+++ VimeoDotNet/VimeoDotNet/Models/Space.cs
@@ -7,6 +7,6 @@ namespace VimeoDotNet.Models
     {
         public long max { get; set; }
         public long free { get; set; }
-        public int used { get; set; }
+        public long used { get; set; }
     }
 }
